set the value of a bit to a given index

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -8,7 +8,7 @@
 * Return: I if it worked else -1
 */
 
-int get_bit(unsigned int n, unsigned int index)
+int get_bit(unsigned long int n, unsigned int index)
 {
 	unsigned int i;
